requirements.txt
aniso8601==4.0.1
Click==7.0
dj-database-url==0.5.0
Django==2.1.3
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.9.0
Flask==1.0.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
PyJWT==1.4.2
pytz==2018.7
shell==1.0.1
six==1.11.0
SQLAlchemy==1.2.13
Werkzeug==0.14.1
whitenoise==4.1.2
